Splitting out CpsFlowExecutionMemoryTest seems to work.
