+++ pom.xml
@@ -62,7 +62,7 @@
         </pluginRepository>
     </pluginRepositories>
     <properties>
-        <test>CpsFlowExecutionTest</test> <!-- TODO -->
+        <test>CpsFlowExecutionTest,CpsFlowExecutionMemoryTest</test> <!-- TODO -->
         <jenkins.version>2.7.3</jenkins.version>
         <jenkins-test-harness.version>2.28-20170925.184800-1</jenkins-test-harness.version> <!-- TODO https://github.com/jenkinsci/jenkins-test-harness/pull/74 -->
         <no-test-jar>false</no-test-jar>
